apt-build-dep-requirements-dev.txt
python-h5py
python-matplotlib
python-pyside
python-numpy
python-pandas
python-scipy
